config: rename files
